Refactor list rentals
