+++ execute.rb
@@ -4,6 +4,7 @@ require './create_person'
 require './list_books'
 require './create_book'
 require './create_rental'
+require './list_rentals.rb'
 
 class Execute
   def initialize(state)
@@ -13,6 +14,7 @@ class Execute
     @list_books = List_Books.new(@state)
     @create_book = Create_Book.new(@state)
     @create_rental = Create_Rental.new(@state)
+    @list_rentals = List_Rentals.new(@state)
   end
   def execute(user_choice) # rubocop:disable Metrics/CyclomaticComplexity
     case user_choice.to_i
@@ -21,7 +23,7 @@ class Execute
     when 3 then @person.create_person
     when 4 then @create_book.create_book
     when 5 then @create_rental.create_rental
-    when 6 then rentals_for_id
+    when 6 then @list_rentals.rentals_for_id
     when 7 then exit_app
     else p 'Please enter a number between 1 and 7'
     end
